+++ Features/Markers/ContractMarkerManager.cs
@@ -1,0 +1,4 @@
+﻿// ...existing code from Main/ContractMarkerManager.cs...
+
+
+