+++ Features/Minimap/Constants.cs
@@ -1,0 +1,4 @@
+﻿// ...existing code from Helpers/Constants.cs...
+
+
+