+++ Features/Minimap/MapPreferences.cs
@@ -1,0 +1,4 @@
+﻿// ...existing code from Helpers/MapPreferences.cs...
+
+
+